+++ extensions/wrapper/clients/typescript-client/package.json
@@ -44,11 +44,11 @@
     },
     "devDependencies": {
         "@trivago/prettier-plugin-sort-imports": "^4.1.1",
-        "@types/node": "^18.15.11",
+        "@types/node": "^18.19.22",
         "prettier": "^2.8.7",
-        "typescript": "^4.9.3",
-        "vite": "^4.2.0",
-        "vite-plugin-dts": "^2.2.0"
+        "typescript": "^4.9.5",
+        "vite": "^4.5.2",
+        "vite-plugin-dts": "^2.3.0"
     },
     "dependencies": {
         "zod": "^3.22.4"
