+++ extensions/wrapper/clients/typescript-client/src/oauth2/utils/HttpUtils.ts
